backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pytz==2023.3.post1
redis==5.0.1
aiohttp==3.9.1
python-dotenv==1.0.0
emails==0.6
celery==5.3.6
prometheus-client==0.19.0
requests==2.31.0
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2 